+++ src/padone/common/model/Article/FeedbackHandler.java
@@ -11,7 +11,7 @@ import org.apache.tomcat.jdbc.pool.DataSource;
 
 public class FeedbackHandler {
 
-	public boolean newFeedback(DataSource datasource,String author,String articleID,String message) {
+	public boolean newFeedback(DataSource datasource,String userID,String articleID,String message) {
 			Connection con = null;
 			try {
 				con = datasource.getConnection();
@@ -20,7 +20,7 @@ public class FeedbackHandler {
 				
 				Object param = new java.sql.Timestamp(date.getTime());
 				System.out.println(param);
-				 int insert = st.executeUpdate("insert into feedback(feedbackID,articleID,author,updatetime,message) select ifNULL(max(feedbackID+0),0)+1,'" + articleID + "','" + author + "','"+ param + "','" + message + "' FROM feedback");
+				 int insert = st.executeUpdate("insert into feedback(feedbackID,articleID,authorID,updatetime,message) select ifNULL(max(feedbackID+0),0)+1,'" + articleID + "','" + userID + "','"+ param + "','" + message + "' FROM feedback");
 				//int insert = st.executeUpdate("insert into feedback"+"(feedbackID,articleID,author,updatetime,message))select ifNULL(max(feedbackID+0),0)+1,'"+ articleID + "','"+ author + "','" + param + "','" + message+"' FROM feedback");	  
 			    st.close();//關閉st
 			} catch (SQLException e) {
@@ -37,28 +37,16 @@ public class FeedbackHandler {
 		try {
 			con = datasource.getConnection();
 			Statement st = con.createStatement();
-			String author="";
-			ResultSet rs=st.executeQuery("select * from feedback where feedbackID='"+feedbackID+"'");
-			while(rs.next()){
-				author=""+rs.getString("author");
-			}
-			if(userID.equals(author)) {
-				String sql="delete from feedback where feedbackID='"+feedbackID+"'";
-				st.executeUpdate(sql);
-				return true;
-			}
-			else {
-				return false;
-			}
-				
-			
+			String sql="delete from feedback where feedbackID='"+feedbackID+"'"+"and authorID='"+userID+"'";
+			st.executeUpdate(sql);
+			return true;
 		} catch (SQLException e) {
 			System.out.println("PatientInstructionServer newInstruction Exception :" + e.toString());
 			e.printStackTrace();
 		}finally {
 		      if (con!=null) try {con.close();}catch (Exception ignore) {}
 		}
-		return true;
+		return false;
 	}
     public ArrayList<Feedback> getFeedback(DataSource datasource,String articleID){
     	Connection con = null;
@@ -69,11 +57,11 @@ public class FeedbackHandler {
 			ResultSet rs=st.executeQuery("select * from feedback where articleID='"+articleID+"'");
 			while(rs.next()) {
 				Feedback temp=new Feedback();
-				String author=rs.getString("author");
+				String authorID=rs.getString("authorID");
 				String id=rs.getString("feedbackID");
 				String updateTime=rs.getString("updateTime");
 				String message=rs.getString("message");
-				temp.feedback(articleID, id, author, message, updateTime);
+				temp.feedback(articleID, id, author,authorID, message, updateTime);
 				result.add(temp);
